requirements.txt
flask==2.0.1
werkzeug==2.0.3
python-dotenv==0.19.0
supabase==1.0.3
flask-cors==3.0.10 